TRIPLEX-851 LightBox add sizes
