+++ src/components/LightBox/enums.ts
@@ -1,0 +1,6 @@
+export enum ELightBoxSize {
+    SM = "sm",
+    MD = "md",
+    LG = "lg",
+    XL = "xl",
+}
