+++ src/components/LightBox/index.ts
@@ -1,2 +1,3 @@
 export * from "./LightBox";
+export * from "./enums";
 export * from "./LightBoxSideOverlay";
