+++ stories/release-notes/v1/1.25.0.mdx
@@ -6,6 +6,6 @@ import { Meta, Title, Heading } from "@storybook/addon-docs/blocks";
 
 <Heading>Изменения</Heading>
 
-**Title**
+**LightBox**
 
-- Description.
+- Добавлен prop `size` (`ELightBoxSize`) для управления шириной контента. Размер контента автоматически применяется к `LightBox.Content`, `LightBox.SideOverlay`, `LightBox.LeftSidebar` и `LightBox.RightSidebar`.
